requirements.txt
openai-whisper>=20231117
torch>=1.9.0
torchaudio>=0.9.0
numpy>=1.21.0
pathlib2>=2.3.7
